Excludes botania blocks from bonemeal and grass stomping

+++ src/main/java/lilylicious/mysteriousmiscellany/enchantment/EnchantGrassStomper.java
@@ -45,7 +45,8 @@ public class EnchantGrassStomper extends EnchantmentBase {
                     target = player.getEntityWorld().getBlockState(blockPos).getBlock();
 
                     if (target instanceof BlockTallGrass
-                            || target instanceof BlockDoublePlant){
+                            || target instanceof BlockDoublePlant
+                            && !target.toString().toLowerCase().contains("botania"))
                         player.getEntityWorld().destroyBlock(blockPos, false);
                     }
                 }
